requirements.txt
Django==3.0.6
djangorestframework==3.11.0
psycopg2-binary==2.8.5
demjson~=2.2.4
pika~=1.1.0
python-dateutil~=2.8.1
django-filter~=2.2.0
